fix: compile error with duplicated enum source members
Goverter creates a case in enum switch cases for every enum constant.
When having multiple enum constants with the same value, goverter
created a compile error because Go doesn't allow multiple cases with the
same constant value.

This commit prevents duplicated switch case statements by either
skipping a constant if the mapping is identical to another one, or by
erroring and having the user to create an explicit mapping.